🚚[Rename] user/ui/component 관심사별로 컴포넌트 폴더 분리
