+++ src/components/features/user/ui/component/edit/ProfileAvatarEditor/ProfileAvatarEditor.stories.tsx
